# talde1erronka2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/tratamenduak": {
            "get": {
                "tags": [
                    "Tratamenduak"
                ],
                "description": "Tratamendu guztiak bueltatzen ditu.",
                "operationId": "970e605f51f4ec20b4721b2ffffa3815",
                "responses": {
                    "200": {
                        "description": "Tratamendu guztiak bueltatu ditu."
                    },
                    "404": {
                        "description": "Ez du tratamendurik topatu."
                    }
                }
            },
            "put": {
                "tags": [
                    "Tratamenduak"
                ],
                "description": "Adierazitako informazioaren arabera tratamendu bat eguneratzen du datubasean.",
                "operationId": "056755cd1a1e96820c39fd8feb6321ad",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Tratamenduaren ID-a",
                                        "type": "integer"
                                    },
                                    "izena": {
                                        "description": "Tratamenduaren izena",
                                        "type": "string"
                                    },
                                    "etxeko_prezioa": {
                                        "description": "Etxeko prezioa",
                                        "type": "integer"
                                    },
                                    "kanpoko_prezioa": {
                                        "description": "Kanpoko prezioa",
                                        "type": "integer"
                                    },
                                    "id_katTratamendu": {
                                        "description": "Tratamenduaren kategoria ID-a",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Tratamendua ondo eguneratu da datubasean."
                    },
                    "404": {
                        "description": "Ez du tratamendurik topatu ID horrekin."
                    }
                }
            },
            "post": {
                "tags": [
                    "Tratamenduak"
                ],
                "description": "Adierazitako informazioarekin tratamendu berri bat txertatzen du datubasean.",
                "operationId": "09cc1369db599165994c72b1edeccce4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "izena": {
                                        "description": "Tratamenduaren izena",
                                        "type": "string"
                                    },
                                    "etxeko_prezioa": {
                                        "description": "Etxeko prezioa",
                                        "type": "integer"
                                    },
                                    "kanpoko_prezioa": {
                                        "description": "Kanpoko prezioa",
                                        "type": "integer"
                                    },
                                    "id_katTratamendu": {
                                        "description": "Tratamenduaren kategoria ID-a",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tratamendua ondo txertatu da datubasean."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tratamenduak"
                ],
                "description": "Adierazitako ID-a erabilita datubasean tratamenduen ezabatze logikoa egiten du.",
                "operationId": "673c8111bbffd29ba83a7eb16113154f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Ezabatzeko tratamenduaren ID-a",
                                        "type": "integer"
                                    },
                                    "ezabatze_data": {
                                        "description": "Ezabatze-data eta ordua",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tratamenduak ondo ezabatu dira datubasean."
                    },
                    "404": {
                        "description": "Ez dago tratamendurik ID horrekin."
                    }
                }
            }
        },
        "/tratamenduByLangile/{group}": {
            "get": {
                "tags": [
                    "Tratamenduak"
                ],
                "description": "Adierazitako ID-aren arabera langileak bueltatzen bueltatzen ditu tratamendutan bereizita.",
                "operationId": "bfe6c94b894c21889e4fef3c7b33aa9a",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Taldearen ID-a",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aukeratutako guztia bueltatzen du."
                    },
                    "404": {
                        "description": "Ez du ezer topatu ID horrekin."
                    }
                }
            }
        },
        "/txanda": {
            "get": {
                "tags": [
                    "Txandak"
                ],
                "description": "Txanda guztiak bueltatzen ditu.",
                "operationId": "93213e0013188279e37b439279635969",
                "responses": {
                    "200": {
                        "description": "Txanda guztiak bueltatu ditu."
                    },
                    "404": {
                        "description": "Ez du txandarik topatu."
                    }
                }
            },
            "post": {
                "tags": [
                    "Txandak"
                ],
                "description": "Adierazitako informazioarekin produktu berri bat txertatzen du datubasean.",
                "operationId": "2f7a39cb51e574c30c940fc96ef062fe",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "izena": {
                                        "description": "Produktuaren izena",
                                        "type": "string"
                                    },
                                    "deskribapena": {
                                        "description": "Produktuaren deskribapena",
                                        "type": "string"
                                    },
                                    "id_kategoria": {
                                        "description": "Kategoriaren ID-a",
                                        "type": "integer"
                                    },
                                    "marka": {
                                        "description": "Produktuaren marka",
                                        "type": "string"
                                    },
                                    "stock": {
                                        "description": "Hasierako stock-a",
                                        "type": "integer"
                                    },
                                    "stock_alerta": {
                                        "description": "Stock-a alertaren kantitatea",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produktua ondo txertatu da datubasean."
                    }
                }
            }
        },
        "/txanda/{id}": {
            "get": {
                "tags": [
                    "Txandak"
                ],
                "description": "Adierazitako ID-aren arabera txanda bat bueltatzen du.",
                "operationId": "b1199555e4a600d851219540e50e573d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Txandaren ID-a",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aukeratutako txanda bueltatzen du."
                    },
                    "404": {
                        "description": "Ez du txandarik topatu ID horrekin."
                    }
                }
            }
        }
    },
...
}